chore: rename gallery to image gallery
This commit renames the `./Gallery/Gallery.tx` component to
`ImageGallery.tsx`. This makes it easier to determine what this
component does at a glance. The commit also removes an unused
react-imgix package import.
