+++ src/components/Gallery/ImageGallery.css
